+++ Scripts/releases.md
@@ -1,18 +1,26 @@
 # mods Scripts Release Notes
 
-<b>Last Update: 2026 Jan 20 [rwp/osu]</b>
+<b>Last Update: 2026 Jan 25 [rwp/osu]</b>
+
+## 2026 Jan 20: modsStatus.pl and ccdStatus scripts
+ * Updated `modsStatus.pl` to version 2 to include `azcam-mods` and `modsCCD` status, cleaned up display
+ * Added `ccdStatus` csh script to go on the Archon servers to provide remote status info for `azcam-mods` and `modsCCD`
+
 
 ## 2026 Jan 20: go_mods script
  * all-in-one script for access to MODS servers, 1 script to replace 6
+
 
 ## 2026 Jan 18: updated modsAlign and modsDisp
  * `modsAlign` updated to use the azcam MEF format for the Archons, advanced version to 3.1.x
  * `modsDisp` updated to use the azcam MEF format for the Archons, version 3.1.0
 
+
 ## 2026 Jan 16: new scripts
  * `start_ccd` is an experimental script to start `azcam-mods` and `modsCCD` in the correct order.
  * `updateSlits` is an updated version of a script of the same name used by instrument support personnel after mask changes to commit the slitmask tables to the running MMC servers.  This version does both MODS at once, and can be run from `robs` or `obs` machines wihtout having to login to the MODS servers proper.
  * cleaned up internal comments to better match the code logic
+
 
 ## 2025 Dec 31: new and modified scripts
  * `stop_ccd` a convenience script to stop `modsCCD` and `azcam-mods` gracefully in a `ccdTerm` tmux session.